SciChart.iOS v2.1.0.1890

+++ v2.x/Examples/SciChartDemo/SciChartDemo/OnboardingViewController.m
@@ -11,7 +11,7 @@
 @implementation OnboardingViewController {
     
     NSMutableArray * viewControllers;
-    int viewsQuantity;
+    NSUInteger viewsQuantity;
     
 }
